pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-notes-enricher"
version = "0.0.1"
description = "Enriquece notas médicas (Markdown/Obsidian) com imagens de múltiplas fontes."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "Pillow>=10.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
pdf = [
  "pdfplumber>=0.11",
  "pdf2image>=1.17",
  "pytesseract>=0.3.10",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[project.scripts]
enricher = "enricher.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
